move global error handler to own file
