+++ app.js
@@ -2,6 +2,7 @@ const {getClient} = require('./config/client.js');
 const {collectCommands} = require('./config/collectors');
 const {extendMutes} = require('./handlers/channelHandlers.js');
 const {applyMute, createMutedRole} = require('./handlers/guildHandlers.js');
+const {unhandledRejectionHandler} = require('./handlers/errorHandlers');
 const {
   messageHandler,
   logDeletedMessages,
@@ -31,20 +32,7 @@ client.on('messageCreate', messageHandler);
 
 client.on('messageDelete', logDeletedMessages);
 
-process.on('unhandledRejection', async (error) => {
-  try {
-    console.error('Unhandled promise rejection:', error);
-    const channel = client.channels.cache.find(
-      (channel) => channel.name === 'audit-logs'
-    );
-    if (channel) {
-      await channel.send(
-        `UnhandledRejection error. Check logs for more info. Type: ${error.name} Message: ${error.message}`
-      );
-    }
-  } catch (rejectionHandlerErr) {
-    console.error('Error in unhandledRejection handler:', rejectionHandlerErr);
-  }
-});
+// Handles unhandled promise rejections which terminate the node process since node v15.
+process.on('unhandledRejection', unhandledRejectionHandler);
 
 client.login(process.env.DISCORD_SECRET_KEY);
